Several updates on UI, more in description
-Shop is removed as discussed in the meeting.
-Everything within the app follows the app.theme.
-Drawer added to switch between screens.
-All screens are now improved aesthetically and functionally.
-Setting and edit profile screens are added.

Special notes: I will try to focus on back-end since not having functions feel too restrictive, and just developing the UI is not a healthy approach for the development.

It would be great if you can also use the app theme file for your parts! I based it around the colors you choose. You can edit your code according to it.